refactor tests to match updated module structure
