+++ bids/layout/tests/test_layout.py
@@ -3,22 +3,19 @@ BIDSLayout class."""
 
 import os
 import re
-from os.path import join, abspath, basename, dirname
+from os.path import join, abspath, basename
 
 import numpy as np
 import pytest
 
-import bids
-from bids.layout import (BIDSLayout, parse_file_entities, add_config_paths,
-                         Query)
+from bids.layout import BIDSLayout, Query
 from bids.layout.index import BIDSLayoutIndexer
-from bids.layout.models import Entity, Config
+from bids.layout.models import Config
 from bids.tests import get_test_data_path
 from bids.utils import natural_sort
 
 from bids.exceptions import (
     BIDSValidationError,
-    ConfigError,
     NoMatchError,
     TargetError,
 )
@@ -487,38 +484,6 @@ def test_to_df(layout_ds117):
                 set(df.columns))
 
 
-@pytest.mark.parametrize("extension_initial_dot", (True, False))
-def test_parse_file_entities(mock_config, extension_initial_dot):
-    filename = '/sub-03_ses-07_run-4_desc-bleargh_sekret.nii.gz'
-
-    dot = '.' if extension_initial_dot else ''
-
-    # Test with entities taken from bids config
-    target = {'subject': '03', 'session': '07', 'run': 4, 'suffix': 'sekret',
-              'extension': dot + 'nii.gz'}
-    assert target == parse_file_entities(filename, config='bids')
-    config = Config.load('bids')
-    assert target == parse_file_entities(filename, config=[config])
-
-    # Test with entities taken from bids and derivatives config
-    target = {'subject': '03', 'session': '07', 'run': 4, 'suffix': 'sekret',
-              'desc': 'bleargh', 'extension': dot + 'nii.gz'}
-    assert target == parse_file_entities(filename)
-    assert target == parse_file_entities(
-        filename, config=['bids', 'derivatives'])
-
-    # Test with list of Entities
-    entities = [
-        Entity('subject', "[/\\\\]sub-([a-zA-Z0-9]+)"),
-        Entity('run', "[_/\\\\]run-0*(\\d+)", dtype=int),
-        Entity('suffix', "[._]*([a-zA-Z0-9]*?)\\.[^/\\\\]+$"),
-        Entity('desc', "desc-([a-zA-Z0-9]+)"),
-    ]
-    # Leave out session to distinguish from previous test target
-    target = {'subject': '03', 'run': 4, 'suffix': 'sekret', 'desc': 'bleargh'}
-    assert target == parse_file_entities(filename, entities=entities)
-
-
 # XXX 0.14: Add dot to extension (difficult to parametrize with module-scoped fixture)
 def test_parse_file_entities_from_layout(layout_synthetic):
     layout = layout_synthetic
@@ -559,20 +524,6 @@ def test_deriv_indexing():
     assert layout.get(scope='derivatives')
     assert layout.get(scope='events')
     assert not layout.get(scope='nonexistent')
-
-
-def test_add_config_paths():
-    bids_dir = dirname(bids.__file__)
-    bids_json = os.path.join(bids_dir, 'layout', 'config', 'bids.json')
-    with pytest.raises(ConfigError) as exc:
-        add_config_paths(test_config1='nonexistentpath.json')
-    assert str(exc.value).startswith('Configuration file')
-    with pytest.raises(ConfigError) as exc:
-        add_config_paths(bids=bids_json)
-    assert str(exc.value).startswith("Configuration 'bids' already")
-    add_config_paths(dummy=bids_json)
-    config = Config.load('dummy')
-    assert 'subject' in config.entities
 
 
 def test_layout_in_scope(layout_ds005, layout_ds005_derivs):
